Enable doors to open depending on mass of objects placed on them
Set the weight of the table and chair, that when combined, will be more than the hard coded threshold to allow the door to open.  Enable overlap events on both the table and the chair.  The player must now stack both the table and chair on the hidden pressure plate to successfully escape the room.