Fix bug where Default Pawn was rolling
When collisions were turned on for the Default Pawn in the last commit, this caused the Default Pawn to then roll around in the world when interacting with other objects in the room.  This issue was solved by locking the X, Y, and Z rotation constraints in the Collision Component attached to the Default Pawn.