Add a new broadcast event to close the doors
This new event revereses the animation of the Open Door movement when there is no longer enough weight on the pressure plate.  This should really be event driven, but is out of the scope of the program, so we will live with the program polling the TriggerVolume every tick.  Minor refactoring also performed.